Disable gfx.canvas.azure.accelerated
According to lsalzman, gfx.canvas.azure.accelerated is now set to false
on nearly all platforms (currently Mac ESR being the only exception and
they want to turn it off there as well). Switching this on causes us to
miss the non-accelerated code paths, so we might miss bugs. Their
recommendation is to turn it off.